+++ lib/src/models/transcript_list.dart
@@ -49,8 +49,9 @@ class TranscriptList extends IterableBase<Transcript> {
     throw NoTranscriptFoundException(
       videoId: videoId,
       requestedLanguages: languages,
-      availableLanguages:
-          transcripts.map((t) => t.languageCode).toList(growable: false),
+      availableLanguages: transcripts
+          .map((t) => t.languageCode)
+          .toList(growable: false),
     );
   }