+++ lib/src/parsing/transcript_list_parser.dart
@@ -88,7 +88,8 @@ class TranscriptListParser {
     required Future<FetchedTranscript> Function(
       String url,
       bool preserveFormatting,
-    ) fetchFunction,
+    )
+    fetchFunction,
   }) {
     try {
       // captionsJson IS the playerCaptionsTracklistRenderer